Refatoração e separação do formulário de produto
