--- src/styles/pages/admin/product/New.module.scss
@@ -1,20 +1,0 @@
-.container {
-  max-width: 40rem;
-
-  margin-block: var(--size-6);
-}
-
-.formTitle {
-  font-size: var(--fs-medium-title);
-
-  margin-bottom: var(--size-3);
-}
-
-.formFields {
-  display: grid;
-  gap: var(--size-3);
-}
-
-.fileInput {
-  width: 100%;
-}
